+++ src/components/index.ts
@@ -2,3 +2,6 @@ export { default as About } from "@/components/About";
 export { default as Anchor } from "@/components/Anchor";
 export { default as Expirience } from "@/components/Expirience";
 export { default as Header } from "@/components/Header";
+export { default as Project } from "@/components/Project";
+
+export * as Card from "@/components/Card";
